@@ -1,91 +1,92 @@
 {
   "name": "@angular-devkit/build-angular",
-  "version": "16.2.12",
+  "version": "17.3.5",
   "description": "Angular Webpack Build Facade",
   "main": "src/index.js",
   "typings": "src/index.d.ts",
   "builders": "builders.json",
   "dependencies": {
-    "@ampproject/remapping": "2.2.1",
-    "@angular-devkit/architect": "0.1602.12",
-    "@angular-devkit/build-webpack": "0.1602.12",
-    "@angular-devkit/core": "16.2.12",
-    "@babel/core": "7.22.9",
-    "@babel/generator": "7.22.9",
+    "@ampproject/remapping": "2.3.0",
+    "@angular-devkit/architect": "0.1703.5",
+    "@angular-devkit/build-webpack": "0.1703.5",
+    "@angular-devkit/core": "17.3.5",
+    "@babel/core": "7.24.0",
+    "@babel/generator": "7.23.6",
     "@babel/helper-annotate-as-pure": "7.22.5",
     "@babel/helper-split-export-declaration": "7.22.6",
-    "@babel/plugin-proposal-async-generator-functions": "7.20.7",
-    "@babel/plugin-transform-async-to-generator": "7.22.5",
-    "@babel/plugin-transform-runtime": "7.22.9",
-    "@babel/preset-env": "7.22.9",
-    "@babel/runtime": "7.22.6",
-    "@babel/template": "7.22.5",
+    "@babel/plugin-transform-async-generator-functions": "7.23.9",
+    "@babel/plugin-transform-async-to-generator": "7.23.3",
+    "@babel/plugin-transform-runtime": "7.24.0",
+    "@babel/preset-env": "7.24.0",
+    "@babel/runtime": "7.24.0",
     "@discoveryjs/json-ext": "0.5.7",
-    "@ngtools/webpack": "16.2.12",
-    "@vitejs/plugin-basic-ssl": "1.0.1",
+    "@ngtools/webpack": "17.3.5",
+    "@vitejs/plugin-basic-ssl": "1.1.0",
     "ansi-colors": "4.1.3",
-    "autoprefixer": "10.4.14",
+    "autoprefixer": "10.4.18",
     "babel-loader": "9.1.3",
     "babel-plugin-istanbul": "6.1.1",
     "browserslist": "^4.21.5",
-    "chokidar": "3.5.3",
     "copy-webpack-plugin": "11.0.0",
-    "critters": "0.0.20",
-    "css-loader": "6.8.1",
-    "esbuild-wasm": "0.18.17",
-    "fast-glob": "3.3.1",
-    "guess-parser": "0.4.22",
-    "https-proxy-agent": "5.0.1",
-    "inquirer": "8.2.4",
-    "jsonc-parser": "3.2.0",
+    "critters": "0.0.22",
+    "css-loader": "6.10.0",
+    "esbuild-wasm": "0.20.1",
+    "fast-glob": "3.3.2",
+    "https-proxy-agent": "7.0.4",
+    "http-proxy-middleware": "2.0.6",
+    "inquirer": "9.2.15",
+    "jsonc-parser": "3.2.1",
     "karma-source-map-support": "1.4.0",
-    "less": "4.1.3",
+    "less": "4.2.0",
     "less-loader": "11.1.0",
     "license-webpack-plugin": "4.0.2",
     "loader-utils": "3.2.1",
-    "magic-string": "0.30.1",
-    "mini-css-extract-plugin": "2.7.6",
-    "mrmime": "1.0.1",
+    "magic-string": "0.30.8",
+    "mini-css-extract-plugin": "2.8.1",
+    "mrmime": "2.0.0",
     "open": "8.4.2",
     "ora": "5.4.1",
     "parse5-html-rewriting-stream": "7.0.0",
-    "picomatch": "2.3.1",
-    "piscina": "4.0.0",
-    "postcss": "8.4.31",
-    "postcss-loader": "7.3.3",
+    "picomatch": "4.0.1",
+    "piscina": "4.4.0",
+    "postcss": "8.4.35",
+    "postcss-loader": "8.1.1",
     "resolve-url-loader": "5.0.0",
     "rxjs": "7.8.1",
-    "sass": "1.64.1",
-    "sass-loader": "13.3.2",
-    "semver": "7.5.4",
-    "source-map-loader": "4.0.1",
+    "sass": "1.71.1",
+    "sass-loader": "14.1.1",
+    "semver": "7.6.0",
+    "source-map-loader": "5.0.0",
     "source-map-support": "0.5.21",
-    "terser": "5.19.2",
-    "text-table": "0.2.0",
+    "terser": "5.29.1",
     "tree-kill": "1.2.2",
-    "tslib": "2.6.1",
-    "vite": "4.5.2",
-    "webpack": "5.88.2",
-    "webpack-dev-middleware": "6.1.1",
+    "tslib": "2.6.2",
+    "undici": "6.11.1",
+    "vite": "5.1.7",
+    "watchpack": "2.4.0",
+    "webpack": "5.90.3",
+    "webpack-dev-middleware": "6.1.2",
     "webpack-dev-server": "4.15.1",
-    "webpack-merge": "5.9.0",
+    "webpack-merge": "5.10.0",
     "webpack-subresource-integrity": "5.1.0"
   },
   "optionalDependencies": {
-    "esbuild": "0.18.17"
+    "esbuild": "0.20.1"
   },
   "peerDependencies": {
-    "@angular/compiler-cli": "^16.0.0",
-    "@angular/localize": "^16.0.0",
-    "@angular/platform-server": "^16.0.0",
-    "@angular/service-worker": "^16.0.0",
+    "@angular/compiler-cli": "^17.0.0",
+    "@angular/localize": "^17.0.0",
+    "@angular/platform-server": "^17.0.0",
+    "@angular/service-worker": "^17.0.0",
+    "@web/test-runner": "^0.18.0",
+    "browser-sync": "^3.0.2",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
     "karma": "^6.3.0",
-    "ng-packagr": "^16.0.0",
+    "ng-packagr": "^17.0.0",
     "protractor": "^7.0.0",
     "tailwindcss": "^2.0.0 || ^3.0.0",
-    "typescript": ">=4.9.3 <5.2"
+    "typescript": ">=5.2 <5.5"
   },
   "peerDependenciesMeta": {
     "@angular/localize": {
@@ -128,7 +135,7 @@
     "url": "https://github.com/angular/angular-cli.git"
   },
   "engines": {
-    "node": "^16.14.0 || >=18.10.0",
+    "node": "^18.13.0 || >=20.9.0",
     "npm": "^6.11.0 || ^7.5.6 || >=8.0.0",
     "yarn": ">= 1.13.0"
   },
