@@ -1,13 +1,22 @@
 {
   "name": "@angular-devkit/build-webpack",
-  "version": "0.1602.12",
+  "version": "0.1703.5",
   "description": "Webpack Builder for Architect",
   "experimental": true,
   "main": "src/index.js",
   "typings": "src/index.d.ts",
   "builders": "builders.json",
+  "exports": {
+    ".": {
+      "types": "./src/index.d.ts",
+      "default": "./src/index.js"
+    },
+    "./package.json": "./package.json",
+    "./*": "./*.js",
+    "./*.js": "./*.js"
+  },
   "dependencies": {
-    "@angular-devkit/architect": "0.1602.12",
+    "@angular-devkit/architect": "0.1703.5",
     "rxjs": "7.8.1"
   },
   "peerDependencies": {
@@ -26,7 +35,7 @@
     "url": "https://github.com/angular/angular-cli.git"
   },
   "engines": {
-    "node": "^16.14.0 || >=18.10.0",
+    "node": "^18.13.0 || >=20.9.0",
     "npm": "^6.11.0 || ^7.5.6 || >=8.0.0",
     "yarn": ">= 1.13.0"
   },
