@@ -1,6 +1,6 @@
 {
   "name": "@angular-devkit/core",
-  "version": "16.2.12",
+  "version": "17.3.5",
   "description": "Angular DevKit - Core Utility Library",
   "main": "src/index.js",
   "typings": "src/index.d.ts",
@@ -32,8 +32,8 @@
   "dependencies": {
     "ajv-formats": "2.1.1",
     "ajv": "8.12.0",
-    "jsonc-parser": "3.2.0",
-    "picomatch": "2.3.1",
+    "jsonc-parser": "3.2.1",
+    "picomatch": "4.0.1",
     "rxjs": "7.8.1",
     "source-map": "0.7.4"
   },
@@ -50,7 +50,7 @@
     "url": "https://github.com/angular/angular-cli.git"
   },
   "engines": {
-    "node": "^16.14.0 || >=18.10.0",
+    "node": "^18.13.0 || >=20.9.0",
     "npm": "^6.11.0 || ^7.5.6 || >=8.0.0",
     "yarn": ">= 1.13.0"
   },
