@@ -1,6 +1,6 @@
 {
   "name": "@angular-devkit/schematics",
-  "version": "16.2.12",
+  "version": "17.3.5",
   "description": "Angular Schematics - Library",
   "main": "src/index.js",
   "typings": "src/index.d.ts",
@@ -18,9 +18,9 @@
     "tooling"
   ],
   "dependencies": {
-    "@angular-devkit/core": "16.2.12",
-    "jsonc-parser": "3.2.0",
-    "magic-string": "0.30.1",
+    "@angular-devkit/core": "17.3.5",
+    "jsonc-parser": "3.2.1",
+    "magic-string": "0.30.8",
     "ora": "5.4.1",
     "rxjs": "7.8.1"
   },
@@ -29,7 +29,7 @@
     "url": "https://github.com/angular/angular-cli.git"
   },
   "engines": {
-    "node": "^16.14.0 || >=18.10.0",
+    "node": "^18.13.0 || >=20.9.0",
     "npm": "^6.11.0 || ^7.5.6 || >=8.0.0",
     "yarn": ">= 1.13.0"
   },
