@@ -1,17 +1,17 @@
 {
   "name": "@angular/animations",
-  "version": "16.2.12",
+  "version": "17.3.5",
   "description": "Angular - animations integration with web-animations",
   "author": "angular",
   "license": "MIT",
   "engines": {
-    "node": "^16.14.0 || >=18.10.0"
+    "node": "^18.13.0 || >=20.9.0"
   },
   "dependencies": {
     "tslib": "^2.3.0"
   },
   "peerDependencies": {
-    "@angular/core": "16.2.12"
+    "@angular/core": "17.3.5"
   },
   "repository": {
     "type": "git",
