@@ -1,6 +1,6 @@
 {
   "name": "@angular/cli",
-  "version": "16.2.12",
+  "version": "17.3.5",
   "description": "CLI tool for Angular",
   "main": "lib/cli/index.js",
   "bin": {
@@ -25,38 +25,39 @@
   },
   "homepage": "https://github.com/angular/angular-cli",
   "dependencies": {
-    "@angular-devkit/architect": "0.1602.12",
-    "@angular-devkit/core": "16.2.12",
-    "@angular-devkit/schematics": "16.2.12",
-    "@schematics/angular": "16.2.12",
+    "@angular-devkit/architect": "0.1703.5",
+    "@angular-devkit/core": "17.3.5",
+    "@angular-devkit/schematics": "17.3.5",
+    "@schematics/angular": "17.3.5",
     "@yarnpkg/lockfile": "1.1.0",
     "ansi-colors": "4.1.3",
-    "ini": "4.1.1",
-    "inquirer": "8.2.4",
-    "jsonc-parser": "3.2.0",
-    "npm-package-arg": "10.1.0",
-    "npm-pick-manifest": "8.0.1",
+    "ini": "4.1.2",
+    "inquirer": "9.2.15",
+    "jsonc-parser": "3.2.1",
+    "npm-package-arg": "11.0.1",
+    "npm-pick-manifest": "9.0.0",
     "open": "8.4.2",
     "ora": "5.4.1",
-    "pacote": "15.2.0",
-    "resolve": "1.22.2",
-    "semver": "7.5.4",
+    "pacote": "17.0.6",
+    "resolve": "1.22.8",
+    "semver": "7.6.0",
     "symbol-observable": "4.0.0",
     "yargs": "17.7.2"
   },
   "ng-update": {
     "migrations": "@schematics/angular/migrations/migration-collection.json",
     "packageGroup": {
-      "@angular/cli": "16.2.12",
-      "@angular-devkit/architect": "0.1602.12",
-      "@angular-devkit/build-angular": "16.2.12",
-      "@angular-devkit/build-webpack": "0.1602.12",
-      "@angular-devkit/core": "16.2.12",
-      "@angular-devkit/schematics": "16.2.12"
+      "@angular/cli": "17.3.5",
+      "@angular/ssr": "17.3.5",
+      "@angular-devkit/architect": "0.1703.5",
+      "@angular-devkit/build-angular": "17.3.5",
+      "@angular-devkit/build-webpack": "0.1703.5",
+      "@angular-devkit/core": "17.3.5",
+      "@angular-devkit/schematics": "17.3.5"
     }
   },
   "engines": {
-    "node": "^16.14.0 || >=18.10.0",
+    "node": "^18.13.0 || >=20.9.0",
     "npm": "^6.11.0 || ^7.5.6 || >=8.0.0",
     "yarn": ">= 1.13.0"
   }
