@@ -1,11 +1,11 @@
 {
   "name": "@angular/common",
-  "version": "16.2.12",
+  "version": "17.3.5",
   "description": "Angular - commonly needed directives and services",
   "author": "angular",
   "license": "MIT",
   "engines": {
-    "node": "^16.14.0 || >=18.10.0"
+    "node": "^18.13.0 || >=20.9.0"
   },
   "locales": "locales",
   "dependencies": {
@@ -16,6 +16,7 @@
       "default": "./locales/global/*.js"
     },
     "./locales/*": {
+      "types": "./locales/*.d.ts",
       "default": "./locales/*.mjs"
     },
     "./package.json": {
@@ -53,7 +54,7 @@
     }
   },
   "peerDependencies": {
-    "@angular/core": "16.2.12",
+    "@angular/core": "17.3.5",
     "rxjs": "^6.5.3 || ^7.4.0"
   },
   "repository": {
