@@ -1,6 +1,6 @@
 {
   "name": "@angular/compiler-cli",
-  "version": "16.2.12",
+  "version": "17.3.5",
   "description": "Angular - the compiler CLI for Node.js",
   "typings": "index.d.ts",
   "bin": {
@@ -43,9 +43,9 @@
     }
   },
   "dependencies": {
-    "@babel/core": "7.23.2",
+    "@babel/core": "7.23.9",
     "@jridgewell/sourcemap-codec": "^1.4.14",
-    "reflect-metadata": "^0.1.2",
+    "reflect-metadata": "^0.2.0",
     "chokidar": "^3.0.0",
     "convert-source-map": "^1.5.1",
     "semver": "^7.0.0",
@@ -53,8 +53,8 @@
     "yargs": "^17.2.1"
   },
   "peerDependencies": {
-    "@angular/compiler": "16.2.12",
-    "typescript": ">=4.9.3 <5.2"
+    "@angular/compiler": "17.3.5",
+    "typescript": ">=5.2 <5.5"
   },
   "repository": {
     "type": "git",
@@ -67,7 +67,7 @@
   ],
   "license": "MIT",
   "engines": {
-    "node": "^16.14.0 || >=18.10.0"
+    "node": "^18.13.0 || >=20.9.0"
   },
   "bugs": {
     "url": "https://github.com/angular/angular/issues"
