@@ -1,17 +1,17 @@
 {
   "name": "@angular/compiler",
-  "version": "16.2.12",
+  "version": "17.3.5",
   "description": "Angular - the compiler library",
   "author": "angular",
   "license": "MIT",
   "engines": {
-    "node": "^16.14.0 || >=18.10.0"
+    "node": "^18.13.0 || >=20.9.0"
   },
   "dependencies": {
     "tslib": "^2.3.0"
   },
   "peerDependencies": {
-    "@angular/core": "16.2.12"
+    "@angular/core": "17.3.5"
   },
   "peerDependenciesMeta": {
     "@angular/core": {
@@ -56,12 +56,6 @@
       "esm2022": "./esm2022/index.mjs",
       "esm": "./esm2022/index.mjs",
       "default": "./fesm2022/compiler.mjs"
-    },
-    "./testing": {
-      "types": "./testing/index.d.ts",
-      "esm2022": "./esm2022/testing/testing.mjs",
-      "esm": "./esm2022/testing/testing.mjs",
-      "default": "./fesm2022/testing.mjs"
     }
   }
 }
